+++ client-react/src/ApiHelpers/RuntimeStackService.ts
@@ -11,9 +11,7 @@ import { AppStackOs } from '../models/stacks/app-stacks';
 export default class RuntimeStackService {
   public static getWebAppConfigurationStacks = (stacksOs: AppStackOs) => {
     return sendHttpRequest<WebAppStack[]>({
-      url: `${Url.serviceHost}stacks/webAppStacks?os=${stacksOs}&api-version=${
-        CommonConstants.ApiVersions.stacksApiVersion20200601
-      }&removeHiddenStacks=${!RuntimeStackService._isShowHiddenStackFlagPassed()}`,
+      url: `${Url.serviceHost}stacks/webAppStacks?${RuntimeStackService._getStackUrlParameter(stacksOs)}`,
       method: 'GET',
     }).then(result => {
       const success = result.metadata.success && !!result.data;
@@ -31,9 +29,7 @@ export default class RuntimeStackService {
 
   public static getFunctionAppConfigurationStacks = (stacksOs: AppStackOs) => {
     return sendHttpRequest<FunctionAppStack[]>({
-      url: `${Url.serviceHost}stacks/functionAppStacks?os=${stacksOs}&api-version=${
-        CommonConstants.ApiVersions.stacksApiVersion20200601
-      }&removeHiddenStacks=${!RuntimeStackService._isShowHiddenStackFlagPassed()}`,
+      url: `${Url.serviceHost}stacks/functionAppStacks?${RuntimeStackService._getStackUrlParameter(stacksOs)}`,
       method: 'GET',
     }).then(result => {
       const success = result.metadata.success && !!result.data;
@@ -56,8 +52,19 @@ export default class RuntimeStackService {
     });
   };
 
+  private static _getStackUrlParameter = (stacksOs: AppStackOs) => {
+    return `os=${stacksOs}&api-version=${
+      CommonConstants.ApiVersions.stacksApiVersion20200601
+    }&removeHiddenStacks=${!RuntimeStackService._isShowHiddenStackFlagPassed()}&removeDeprecatedStacks=${!RuntimeStackService._isShowDeprecatedStackFlagPassed()}`;
+  };
+
   private static _isShowHiddenStackFlagPassed = () => {
     const flagValue = Url.getFeatureValue(CommonConstants.FeatureFlags.showHiddenStacks);
+    return flagValue && flagValue.toLocaleLowerCase() === 'true';
+  };
+
+  private static _isShowDeprecatedStackFlagPassed = () => {
+    const flagValue = Url.getFeatureValue(CommonConstants.FeatureFlags.showDeprecatedStacks);
     return flagValue && flagValue.toLocaleLowerCase() === 'true';
   };
 }
